Split CLI interface into smaller files
